renamed the ObstacleMovement script to ObstacleUpdateManager, renamed ObstacleManager to ObstacleSpawnManager, and made it to where the obstacles are destroyed when moving behind the camera
