+++ Assets/Environment/ObstaclePrefabs/IndividualTestObstacle1.prefab
@@ -12,6 +12,7 @@ GameObject:
   - component: {fileID: 6956769021641945020}
   - component: {fileID: 3263295359704392993}
   - component: {fileID: 3722307948315498397}
+  - component: {fileID: -8022089076990304809}
   m_Layer: 0
   m_Name: IndividualTestObstacle1
   m_TagString: Untagged
@@ -105,3 +106,16 @@ BoxCollider:
   serializedVersion: 3
   m_Size: {x: 1, y: 1, z: 1}
   m_Center: {x: 0, y: 0, z: 0}
+--- !u!114 &-8022089076990304809
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6979560732441829179}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 277845b5dcc06514b9f1f5f681d7b4c5, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _obstacleSpeed: 5
