+++ Assets/Environment/ObstaclePrefabs/TestObstacleSet1.prefab
@@ -222,6 +222,7 @@ GameObject:
   - component: {fileID: 2665224920809028844}
   - component: {fileID: 8073528014919394347}
   - component: {fileID: 571640735811243125}
+  - component: {fileID: 8707778085209326021}
   m_Layer: 0
   m_Name: TestObstacleSet1
   m_TagString: Untagged
@@ -317,3 +318,16 @@ BoxCollider:
   serializedVersion: 3
   m_Size: {x: 1, y: 1, z: 1}
   m_Center: {x: 0, y: 0, z: 0}
+--- !u!114 &8707778085209326021
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6753433624405173161}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 277845b5dcc06514b9f1f5f681d7b4c5, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _obstacleSpeed: 5
